Add files via upload
**Brief description of the dataset**

This dataset provides training data on weather from 1st January 2013 to 31st December 2016. An additional testing set is available, enclosing the period 1st January 2017 to 31st March 2017, which corresponds to the forecast interval. There are 4 parameters describing weather, that are *meantemp*, *humidity*, *wind_speed*, and *meanpressure*.

**Project Steps**

The main aim of this project is to perform *weather forecasting* for the period January-March 2017.
Following what has been seen during the lessons, a comprehensive pipeline should be devised, including:

1. Loading, converting and cleaning of the data;
2. Exploring the dataset with descriptive statistics and frequency analysis. Using appropriate graphs to visualise the data at hand;
3. Describing the time series patterns (visually and numerically);
4. Focusing on *meantemp* (temperature as dependent variable), building the most appropriate model(s) to forecast the data for the specified period. In particular, using as test data those contained in the corresponding csv file. Evaluating the model performance using different metrics. Visualising the results with appropriate graphs.

For the different steps, comment on the main results and any relevant observation/finding you have noticed.